HashMap
Its an array of linkedlists. gets a hashcode from the LogObject's tag to find the index. Then creates LinkedList of LogObjects that share the same Tag. The hashing gets you to the start of the node so you dont have to iterate through 10,000 elements to find the elements you need. You only iterate through the tags you are looking for.