HashMap generic
I tried to implement the HashMap as a generic object so it would be applicable to multiple things but I got stuck and couldn't quite get it to work. I'll put more work into it in the future hopefully. I have a non-generic object that will suffice in the meantime.