Fixed generic HashTable
This class takes objects and puts them into a hashtable that organizes data into categories. Each class will provide a way to categorize the element. This category is used to calculate which index the item will go on. On collision: If the object we are trying to put has the same category as the value we collided with, separate chain it to that index. If the categories do not match, linearly probe until we find an empty spot or the category index we are supposed to be at. The purpose is to help datasets that have large element counts for few categories. It allows direct access to elements of a category instead of having to iterate through every element and pick out the elements with the desired property. The only excess travel it needs to do to get a list of all the elements of a certain category is navigate the array if the category was placed at a different index than its hash due to collisions. Simply, in worst case it only needs to navigate all of the categories instead of all the elements.A class that implements the CategoricalHashTableInterface. This interface makes the class have a method available that provides this class with what value to categorize this class with, also a toString for prettier outputs