Categorical Hash Map can now put whole Lists
So when we are parsing equations, we create a list of log objects, it would be redundent to iterate through this list just to have the hash map recreate the same list. It can now take the whole list, and under the assumption that all elements have the same tag it takes the first elements tag and hashes it to place the whole list.