Added Math Channels
There is now a button that launches a dialog where the user can enter an equation. This equation can take variables of tags that already exist in the hashmap, apply basic operations to them using static numbers or other variables to create a new list of LogObjects which is put in the hashmap. So for example there exists a dataset of elements which are coolant as a function of time. The coolant is in farhenheit. I can enter the equation ($(Time,Coolant)-32)*(5/9) to create a new dataset of Time,CoolantInCelcius.