Added Vehicle Data capabilities
The user can now, create, save, and load vehicle dynamic data which is essentially a class that holds a list of variables and an associated values. So the user can save things such as mass, gear ratios, etc to be available in the program so they don't have to reference them from somewhere else later.