Changed File System Processing
The scope of the project has gone beyond the point where a single CSV file is enough for the user to have all the data convinently. Our current approach is to append new data elements to the bottom of the file. So it starts with the CSV portion then comes to the VehicleData section. On approach of the VehicleData section, the while loop breaks and a new while loop begins to handle the VehicleData. If we come to a point where more data is to be saved, it would be appended to the end of the file and another while loop would be added to operate on that data.