Added Functionality to create and break on laps
Displaying markers is broken when trying to view more than one lap at a time. This occurs because the chart moves the second lap (which does not start at 0) over to 0 so that the laps can truly be compared. This means that all the domain of all the log point on the second lap are subtracted to create this. the markers are not moved over. Because if they are moved over then they may be inaccurate for the first lap.