Created a reserved note and lap static markers are always shown on bottom.
The user is no longer allowed to enter "Start/End Lap#" as one of their notes. This is now a reserved note that indicates the static marker was created by the LapBreaker. The static markers created by the LapBreaker are always shown at the bottom because otherwise they clog the list view.