Added method for bounds
There is now a method that calculates the equation but only adds the calculated element to the data set if its between certain bounds, if its not within the bounds the last data set is written. This is effective for mapping items such as brake balance. When the driver is not on the brakes, the torque applied on the rotors is very close to 0 resulting in balances that skyrocket higher or lower than expected. This cuts out those data points.