Added new import functionalities
The user can now import several new imports at a time, each of which are launched on new windows. The user can now select a TXT file to import. This TXT file is in the format the Pi outputs, which is "iso format of time HexStringOfData". This allows for quicker data analysis directly off of the car.