Asks for Post Processing
The data now asks the user if they want to post process. This is so that if the data is imported from the tablet and the data is not synchronized, in which case the post processing would fail, the user can decline to run the other equations and view raw data.