Engine Menu Added!
The Engine Menu currently has two options. One of which is under development and disabled. The other clears the frame and sets up 5 charts for RPM, TPS, Lambda, and Fuel open time. The big chart overlays RPM, TPS, and Lambda. There are 4 smaller charts which show each of the components individually.