Revert "Added a way for the focused chart to be edited by key presses."
This reverts commit d5c6a6718a3c67f8c6f0a418abc7251ef7ce336a.

This also changes the way that the user interacts with the chart. Now every chartFrame has a menu bar. Depending on the focused chart (handled by Java), that menu bar can be used. So the same keystrokes are available on every chart. Alt+KEYS is the default format for editing a chart while CTRL+KEYS is the default for editing the whole application.