Added cut data feature.
Added ability for the user to cut the data of any dataset that is currently on the screen. Potential break points could be if the user deletes and entire dataset. Lots of empty tags floating around. But this is accomplished the same way a lap is created, by selecting two points on a chart. All math channels are regenerated through the EquationEvaluater.